Move RunningApp to running_app.py
