+++ modal/_container_entrypoint.py
@@ -28,13 +28,13 @@ from ._proxy_tunnel import proxy_tunnel
 from ._serialization import deserialize
 from ._utils.async_utils import TaskContext, synchronizer
 from ._utils.function_utils import LocalFunctionError, is_async as get_is_async, is_global_function, method_has_params
-from .app import RunningApp
 from .client import Client, _Client
 from .cls import Cls
 from .config import logger
 from .exception import ExecutionError, InputCancellation, InvalidError
 from .functions import Function, _Function, _set_current_context_ids
 from .partial_function import _find_callables_for_obj, _PartialFunctionFlags
+from .running_app import RunningApp
 from .stub import Stub, _Stub
 
 if TYPE_CHECKING:
